+++ src/types/superAdminUsers.ts
@@ -1,0 +1,10 @@
+export interface UserData {
+  no: string;
+  id: string;
+  name: string;
+  avatar: string;
+  coins: number;
+  plan: string;
+  date: string;
+  isBlocked: boolean;
+}
